refactor: intValue를 Math.toIntExact로 수정
- 오버플로우 발생 시, 예외를 발생하도록 하기 위해

+++ backend/src/main/java/todoktodok/backend/discussion/application/service/query/DiscussionCommentCountDto.java
@@ -12,8 +12,8 @@ public record DiscussionCommentCountDto(
     ) {
         this(
                 discussionId,
-                commentCount.intValue(),
-                replyCount.intValue()
+                Math.toIntExact(commentCount),
+                Math.toIntExact(replyCount)
         );
     }
 }
